Move comparator to utils file
